+++ src/main/java/io/github/astrapi69/gradle/migration/extension/package-info.java
@@ -1,0 +1,17 @@
+/**
+ * This package provides utility classes for working with Gradle project dependencies and Git
+ * repositories. The functionalities include:
+ *
+ * <ul>
+ * <li>Managing and transforming Gradle dependencies, including retrieving dependencies content from
+ * a build.gradle file, converting project names between naming conventions, and generating
+ * dependency-related structures for Gradle projects.</li>
+ * <li>Interacting with Git repositories, such as adding files to Git, unstaging files, and checking
+ * the status of a repository using shell commands.</li>
+ * </ul>
+ *
+ * The classes in this package are designed to facilitate the migration and maintenance of
+ * Gradle-based projects, particularly when dealing with version management and source control
+ * operations.
+ */
+package io.github.astrapi69.gradle.migration.extension;