+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 ## Change log
 ----------------------
 
+Version 2.4
+-------------
+
+ADDED:
+
+- new class GitExtensions that provides methods for git operations to a Git repository
+- new package-info.java files for all packages
+
+
 Version 2.3
 -------------